Refactor Jira models into separate modules for improved organization
- Moved Jira models into distinct modules: agile, comment, common, issue, project, search, workflow, and worklog.
- Updated imports in the main `__init__.py` to reflect the new structure.
- Enhanced maintainability and clarity of the codebase by organizing models by entity type.
- Removed the old monolithic `jira.py` file to streamline the model definitions.
